+++ src/beamlime/kafka/__init__.py
@@ -1,10 +1,11 @@
 # SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
+# Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 
-from .stream_mapping import InputStreamKey, KafkaTopic, StreamMapping
+from .stream_mapping import InputStreamKey, KafkaTopic, StreamLUT, StreamMapping
 
 __all__ = [
     'InputStreamKey',
     'KafkaTopic',
+    'StreamLUT',
     'StreamMapping',
 ]
