+++ tests/services/detector_data_test.py
@@ -10,8 +10,8 @@ from streaming_data_types import eventdata_ev44
 from beamlime import StreamKind
 from beamlime.config.raw_detectors import available_instruments, get_config
 from beamlime.config.streams import stream_kind_to_topic
+from beamlime.core.handler import source_name
 from beamlime.fakes import FakeMessageSink
-from beamlime.kafka.helpers import source_name
 from beamlime.kafka.message_adapter import FakeKafkaMessage, KafkaMessage
 from beamlime.kafka.sink import UnrollingSinkAdapter
 from beamlime.kafka.source import KafkaConsumer
